Hive as builder for Swarm
We want to avoid initializing solutions for the Hive up front, especially in a
single-threaded way. So, we move `working` and `best` into a new struct, which
is generated from a Hive once we've set all the parameters we want to.

The result is that a Hive is now a builder for a Swarm, which actually
maintains the relevant state for running the ABC algorithm.