Observers avoid candidates that are being scouted
Previously, if a candidate expired and a scout began working on it,
another thread could work on the expired solution, then immediately
clobber the scouted solution with a variant on the one that just got
replaced.

To help prevent this, we maintain a set of which candidates are
currently being scouted. Candidates in that set are no longer
eligible for observation.