Switch to context-oriented interface.
In cases where the evaluation process involves something complex, like
communicating with another thread/process/host, it can be helpful or
necessary to have a long-lived object that can handle a demanding
relationship with the outside world.

To serve this need, the library has been reworked so that the Context
trait, which takes over from a Solution::Builder, is the main type
parameter used by the Hive struct.

The Context lasts for the whole life span of the Hive, and is
responsible for initializing, evaluating, and mutating solutions.
Indeed, Solution is now an associated type within a Context.

This is a breaking change, necessitating a minor version bump.