Expose a more detailed Hive API
The goal of this update is to allow the Hive to be used in more
interesting ways, and to enable complex interactions with the Context
by other parts of the program. To this end, the stream method has
been decomposed into a set_sender method (that uses a &mut self) and
run_forever. A context method has been added, which lets other threads
use a reference to the wrapped Context while the hive is running.

The stream() method has been left in place, since it provides a useful
basic interface for the algorithm.