Send current best when setting a sender
By the time a sender is set on a Hive, it has already been initialized,
so it already has a current best result. We now send this over the
sender right away.

This change involves getting a lock on the current best solution, so
Hive.set_sender now returns an abc::Result.