Drop row lock while checking for best in scouting
Previously, in the work_on method, the a worker thread would hold a
write lock on a solution being scouted while examining it as a
candidate for best solution. This is a pretty minor deal - no other
thread should be working on a solution being scouted anyway - but
the code is a little cleaner this way.

Also, blocks are now preferred to explicit drop() statements when
working with mutexes.